+++ CHANGELOG.md
@@ -50,6 +50,9 @@ CLI command and its behaviour. There are no guarantees of stability for the
 
 ### Fixed
 
+- `annotate`'s '`--style` now works again when used for a file with an
+  unrecognised extension. (#909)
+
 ### Security
 
 ## 3.0.1 - 2024-01-19